requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-cas-ng==4.3.0
django-extensions==3.1.5
djangorestframework==3.13.1
django-storages==1.13.2
drf-spectacular==0.27.2
google-api-core==2.11.0
google-api-python-client==2.105.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
# psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.2.0
slack_sdk==3.20.2
